Renamed plugin to Terminal and automatically move cursor to end of terminal when moving through command history.
