+++ plugins/Terminal/css/roboto-mono.css
