+++ plugins/Terminal/css/style.css
