+++ plugins/Terminal/index.html
@@ -12,7 +12,7 @@
 	</head>
 	<body>
         <header class="header">
-            <div class="title">Command Line</div>
+            <div class="title">Terminal</div>
         </header>
 		<!-- React root -->
 		<div id="react-root"></div>
