+++ plugins/Terminal/js/actions/commandline.js
