+++ plugins/Terminal/js/actions/error.js
