+++ plugins/Terminal/js/components/app.js
