+++ plugins/Terminal/js/components/commandhistorylist.js
