+++ plugins/Terminal/js/constants/commandline.js
@@ -5,3 +5,5 @@ export const LOAD_NEXT_COMMAND = 'LOAD_NEXT_COMMAND'
 export const SET_CURRENT_COMMAND = 'SET_CURRENT_COMMAND'
 export const SHOW_WALLET_PROMPT = 'SHOW_WALLET_PROMPT'
 export const HIDE_WALLET_PROMPT = 'HIDE_WALLET_PROMPT'
+export const SHOW_SEED_PROMPT = 'SHOW_SEED_PROMPT'
+export const HIDE_SEED_PROMPT = 'HIDE_SEED_PROMPT'
