+++ plugins/Terminal/js/constants/error.js
