+++ plugins/Terminal/js/containers/commandhistorylist.js
