+++ plugins/Terminal/js/containers/commandinput.js
