+++ plugins/Terminal/js/containers/commandline.js
