+++ plugins/Terminal/js/containers/walletpasswordprompt.js
