+++ plugins/Terminal/js/reducers/commandline.js
