+++ plugins/Terminal/js/reducers/index.js
